+ response parser test that uses the whole code base as a test
(It fails, but gets pretty far so works with most stuff)

Doing this with the thesis that LLMs write markdown easier than a custom format, so must parse that.

But a minor detour from that theme might be to include the path that we are already using as some form of format, at the end of each codeblock as well as the beginning and using that to identify regions.

Any static closing flag would have to be ref counted for the same nesting you deal with in markdown, but a dynamic closing flag like repeating the path, would be much less likely to be repeated in code, and if it were paired with the fuzzy markdown parsing, it might be at least *more* stable, but also still at least follow markdown without having to do anything but give it the examples it already has with the files to context string generation